requirements.txt
argparse
opencv-python
numpy
pycryptodome
